# addons/efishery_module/static/swagger.json
openapi: 3.0.0
info:
 description: >
   API Bertujuan untuk integrasi sale order odoo dengan internal sistem eFishery
 title: eFishery Service ERP Integration
 version: "1.0.0"
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT 
  parameters:
    order_id:
     name: order_id
     description: Id Order
     example: 34576
     in: path
     required: true
     schema:
       type: integer
  schemas:
   order:
     type: object
     properties:
       name:
         type: string
       partner_id:
         type: integer
       date_order:
         type: string
       company_id:
         type: integer       
       order_line:
         type: array
         items:
           type: object
           properties:
             product_id:
               type: integer
             name:
               type: string
             product_uom_qty:
               type: integer
             product_uom:
               type: integer
             price_unit:
               type: integer
  requestBodies:
   order:
     content:
       application/json:
         schema:
           $ref: "#/components/schemas/order"
         example:
           name: "ASD-852/FDR/01/15"
           partner_id: 85555
           date_order: '2021-06-10 06:47:27'
           company_id: 1
           order_line:
           -  product_id: 8626
              product_uom_qty: 1
              product_uom: 24
              price_unit: 40000
security:
  - bearerAuth: []
servers:
  - url: http://localhost:8069/
    description: Odoo Server
paths:
 /order/list:
   get:
    summary: Get quotation order
    description: >
      Get quotation order yang sudah ada
    responses:
      "200":
        description: "Ok"
        content:
          application/json:
            examples:
              response:
                value:
                  success: true
                  message: "Data found"
                  data:
                    name: "ASD-852/FDR/01/15"
                    partner_id: 85555
                    date_order: '2021-06-10 06:47:27'
                    company_id: 1
                    relationship:
                      partner:
                      - partner_id: 85555
                        name: "John"
                        address: "the address"
                      product:
                      - id: 8626
                        name: 'Product'
                        description: 'The Product'
                        price: 23324
                      company: 
                      - id: 1
                        name: 'Company'
                        description: 'The company'
                      uom:
                      - id: 24
                        name: 'Kg'
                        description: 'Satuan berat'
                    order_line:
                    - price_unit: 40000
                      product_id: 8626
                      product_uom_qty: 1
                      product_uom: 24

      "401":
        description: "Unauthorized"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Token Not Found"
      "400":
        description: "Bad request"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "order id not found"
      "500":
        description: "Internal Server Error"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Internal Server Error"
    tags:
      - Sale Order
 /order:
   post:
     summary: Membuat quotation order baru
     description: >
       Membuat quotation order baru
     requestBody:
       $ref: "#/components/requestBodies/order"
     responses:
       "200":
         description: "Ok"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: true
                   message: "Success"
       "401":
         description: "Unauthorized"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   message: "Token Not Found"
       "400":
         description: "Bad request"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   data:
                     name: Required
                     partner_id: Not found
                     date_order: Invalid format date
                     company_id: Required
                     order_line: Is not array
                     order_line.0.product_id: Is not integer
                     order_line.0.product_uom_qty: Is not integer
                     order_line.0.product_uom: Is not integer
                     order_line.0.price_unit: Is not integer
       "409":
         description: "Conflict"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   data:
                     name: Already exists
       "500":
         description: "Internal Server Error"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   message: "Internal Server Error"
     tags:
       - Sale Order
 /order/get/{order_id}:
   get:
    summary: Get quotation order
    description: >
      Get quotation order yang sudah ada
    parameters:
      - $ref: "#/components/parameters/order_id"
    responses:
      "200":
        description: "Ok"
        content:
          application/json:
            examples:
              response:
                value:
                  success: true
                  message: "Data found"
                  data:
                    name: "ASD-852/FDR/01/15"
                    partner_id: 85555
                    date_order: '2021-06-10 06:47:27'
                    company_id: 1
                    relationship:
                      partner:
                      - partner_id: 85555
                        name: "John"
                        address: "the address"
                      product:
                      - id: 8626
                        name: 'Product'
                        description: 'The Product'
                        price: 23324
                      company: 
                      - id: 1
                        name: 'Company'
                        description: 'The company'
                      uom:
                      - id: 24
                        name: 'Kg'
                        description: 'Satuan berat'
                    order_line:
                    - price_unit: 40000
                      product_id: 8626
                      product_uom_qty: 1
                      product_uom: 24

      "401":
        description: "Unauthorized"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Token Not Found"
      "400":
        description: "Bad request"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "order id not found"
      "500":
        description: "Internal Server Error"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Internal Server Error"
    tags:
      - Sale Order
      
 /order/update: 
   put:
     summary: Update quotation order
     description: >
       Update quotation order yang sudah ada
     requestBody:
       $ref: "#/components/requestBodies/order"
     responses:
       "200":
         description: "Ok"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: true
                   message: "Success"
       "401":
         description: "Unauthorized"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   message: "Token Not Found"
       "400":
         description: "Bad request"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   data:
                     name: Required
                     partner_id: Not found
                     date_order: Invalid format date
                     company_id: Required
                     order_line: Is not array
                     order_line.0.sequence: Required
                     order_line.0.product_id: Is not integer
                     order_line.0.product_uom_qty: Is not integer
                     order_line.0.product_uom: Is not integer
                     order_line.0.price_unit: Is not integer
       "409":
         description: "Conflict"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   data:
                     name: Already exists
       "500":
         description: "Internal Server Error"
         content:
           application/json:
             examples:
               response:
                 value:
                   success: false
                   message: "Internal Server Error"

     tags:
       - Sale Order
 /order/delete/{order_id}:
   delete:
    summary: Delete Quotation
    description: >
      Menghapus quotation order
    parameters:
      - $ref: "#/components/parameters/order_id"
    responses:
      "200":
        description: "Ok"
        content:
          application/json:
            examples:
              response:
                value:
                  success: true
                  message: "Delete Success"

      "401":
        description: "Unauthorized"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Token Not Found"
      "400":
        description: "Bad request"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "order id not found"
      "500":
        description: "Internal Server Error"
        content:
          application/json:
            examples:
              response:
                value:
                  success: false
                  message: "Internal Server Error"
    tags:
      - Sale Order
  